Register user management procedures conditionally
Only register user management procedures if native realm is enabled.

- Divide enterprise security procedures into always available `SecurityProcedures`,
and `UserManagementProcedures` that requires a `UserManager`.
