+++ enterprise/security/src/main/java/org/neo4j/server/security/enterprise/auth/SecurityProcedures.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2002-2016 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.server.security.enterprise.auth;
+
+import java.io.IOException;
+import java.util.stream.Stream;
+
+import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
+import org.neo4j.procedure.Description;
+import org.neo4j.procedure.Procedure;
+
+import static org.neo4j.procedure.Mode.DBMS;
+
+@SuppressWarnings( "unused" )
+public class SecurityProcedures extends AuthProceduresBase
+{
+    @Description( "Show the current user." )
+    @Procedure( name = "dbms.security.showCurrentUser", mode = DBMS )
+    public Stream<UserResult> showCurrentUser() throws InvalidArgumentsException, IOException
+    {
+        // TODO: Support LDAP or plugin realms with native realm disabled
+        StandardEnterpriseAuthSubject enterpriseSubject = StandardEnterpriseAuthSubject.castOrFail( authSubject );
+        EnterpriseUserManager userManager = enterpriseSubject.getUserManager();
+        return Stream.of( new UserResult( enterpriseSubject.username(),
+                userManager.getRoleNamesForUser( enterpriseSubject.username() ),
+                userManager.getUser( enterpriseSubject.username() ).getFlags() ) );
+    }
+
+    @Description( "Clears authentication and authorization cache." )
+    @Procedure( name = "dbms.security.clearAuthCache", mode = DBMS )
+    public void clearAuthenticationCache()
+    {
+        ensureAdminAuthSubject().clearAuthCache();
+    }
+}
